+++ paddle/fluid/framework/ir/graph_pattern_detector.cc
@@ -2045,6 +2045,21 @@ PDNode *patterns::MatmulElementwiseAdd::operator()(
   return elementwise_add_out;
 }
 
+void patterns::MulMatmulMatmulV2::operator()(const std::unordered_set<std::string>& ops_type) {
+  auto ops = pattern->NewNode(ops_repr())->assert_is_ops(ops_type);
+  auto ops_x = pattern->NewNode(ops_x_repr())
+                         ->AsInput()
+                         ->assert_is_ops_input(ops_type, "X");
+  auto ops_y = pattern->NewNode(ops_y_repr())
+                         ->AsInput()
+                         ->assert_is_ops_input(ops_type, "Y");
+  auto ops_out = pattern->NewNode(ops_out_repr())
+                        ->AsOutput()
+                        ->assert_is_ops_output(ops_type, "Out");
+
+  ops->LinksFrom({ops_x, ops_y}).LinksTo({ops_out});
+}
+
 PDNode *patterns::ResidualElementwise::operator()(
     PDNode *op_var,
     PDNode *residual_var,
