+++ paddle/fluid/inference/api/analysis_predictor.cc
@@ -2048,7 +2048,6 @@ USE_TRT_CONVERTER(exp);
 USE_TRT_CONVERTER(log);
 USE_TRT_CONVERTER(sigmoid);
 USE_TRT_CONVERTER(tanh);
-USE_TRT_CONVERTER(fc);
 USE_TRT_CONVERTER(pool2d);
 USE_TRT_CONVERTER(softmax);
 USE_TRT_CONVERTER(batch_norm);
