+++ paddle/fluid/inference/tensorrt/plugin/matrix_multiply_op_int8_plugin.cu
@@ -11,7 +11,7 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
-#include "paddle/fluid/inference/tensorrt/plugin/matmul_op_int8_plugin.h"
+#include "paddle/fluid/inference/tensorrt/plugin/matrix_multiply_op_int8_plugin.h"
 
 namespace plf = paddle::platform;
 namespace dyl = paddle::platform::dynload;
